mark

+++ Marksheet/ViewModels/MarksheetVM.cs
@@ -21,24 +21,24 @@ namespace Marksheet.ViewModels
         public string Term { get; set; }
         public int Attendance { get; set; }
 
-        public int EnglishPM { get; set; }//PM-Practical Marks
-        public int EnglishTM { get; set; }//TM-Therory Marks
-        public int NepaliTM { get; set; }
-        public int NepaliPM { get; set; }
-        public int MathTM { get; set; }
-        public int SocialTM { get; set; }
-        public int SocialPM { get; set; }
-        public int HealthPM { get; set; }
-        public int HealthTM { get; set; }
-        public int SciencePM { get; set; }
-        public int ScienceTM { get; set; }
+        public decimal EnglishPM { get; set; }//PM-Practical Marks
+        public decimal EnglishTM { get; set; }//TM-Therory Marks
+        public decimal NepaliTM { get; set; }
+        public decimal NepaliPM { get; set; }
+        public decimal MathTM { get; set; }
+        public decimal SocialTM { get; set; }
+        public decimal SocialPM { get; set; }
+        public decimal HealthPM { get; set; }
+        public decimal HealthTM { get; set; }
+        public decimal SciencePM { get; set; }
+        public decimal ScienceTM { get; set; }
         public string Optional1Name { get; set; }
         public bool Opt1HasPractical { get; set; }
-        public int Optional1PM { get; set; }
-        public int Optional1TM { get; set; }
-        public int ObtePM { get; set; }
-        public int ObteTM { get; set; }
-        public int MoralTM { get; set; }
-        public int MoralPM { get; set; }
+        public decimal Optional1PM { get; set; }
+        public decimal Optional1TM { get; set; }
+        public decimal ObtePM { get; set; }
+        public decimal ObteTM { get; set; }
+        public decimal MoralTM { get; set; }
+        public decimal MoralPM { get; set; }
     }
 }
